requirements-chatterbox-multi-windows.txt
chatterbox-tts==0.1.2

audiotsm==0.1.2
faster-whisper==1.2.0
ffmpeg-python==0.2.0
librosa==0.11.0
loguru==0.7.3
mutagen==1.47.0
natsort==8.4.0
nvidia-ml-py
psutil
pyloudnorm==0.1.1
pysbd==0.3.4
readchar==4.2.1
sounddevice==0.5.2
soundfile==0.13.1
watchdog==6.0.0
xxhash==3.5.0
num2words==0.5.14